Update remains on products in/out/move

+++ src/controllers/products_in.ts
@@ -2,6 +2,7 @@ import { Request, Response } from 'express'
 import ProductIn from '../models/products_in'
 import mongoose from 'mongoose'
 import { IUser } from '../models/user'
+import { changeRemains } from './remains'
 
 export const createProductIn = async (req: Request, res: Response) => {
   try {
@@ -18,6 +19,15 @@ export const createProductIn = async (req: Request, res: Response) => {
       products,
     })
     await new_product_in.save()
+    await changeRemains(
+      products.map(
+        (product: { product: string; name: string; quantity: number }) => ({
+          warehouse,
+          product: product.product,
+          quantity_add: product.quantity,
+        })
+      )
+    )
     res.send('ProductIn created')
   } catch (err: any) {
     res.status(400).send(err.message)
@@ -29,6 +39,15 @@ export const removeProductIn = async (req: Request, res: Response) => {
     const { id } = req.params
     if (!mongoose.Types.ObjectId.isValid(id))
       return res.status(404).send(`No ProductIn with id: ${id}`)
+    const product_in = await ProductIn.findById(id)
+    if (!product_in) throw new Error('ProductIn not found')
+    await changeRemains(
+      product_in.products.map(product => ({
+        warehouse: product_in.warehouse,
+        product: product.product,
+        quantity_add: -product.quantity,
+      }))
+    )
     await ProductIn.findByIdAndRemove(id)
     res.status(200).send('Deleted')
   } catch (err: any) {
@@ -46,13 +65,31 @@ export const updateProductIn = async (req: Request, res: Response) => {
 
     if (!mongoose.Types.ObjectId.isValid(id))
       return res.status(404).send(`No product_in with id: ${id}`)
+
     const old_product_in = await ProductIn.findById(id)
     if (!old_product_in) return res.status(400).send(`ProductIn doens't Exists`)
+
+    await changeRemains(
+      old_product_in.products.map(product => ({
+        warehouse: old_product_in.warehouse,
+        product: product.product,
+        quantity_add: -product.quantity,
+      }))
+    )
 
     await ProductIn.findByIdAndUpdate(
       id,
       { warehouse, date, comment, products, user },
       { new: true }
+    )
+    await changeRemains(
+      products.map(
+        (product: { product: string; name: string; quantity: number }) => ({
+          warehouse,
+          product: product.product,
+          quantity_add: product.quantity,
+        })
+      )
     )
     res.send('ProductIn updated')
   } catch (err: any) {
